requirements.txt
numpy==2.1.1
pillow==10.4.0
tqdm==4.66.5
